Makefile: use ls -1 instead of ls -l
Don't use `ls -l` (dash ell) to count files because on my machine (mac OSX)
that command prints an extra line with the total file size. That line
increases the file count by 1.

Use `ls -1` instead to make sure that every filename appears on a
new line for `wc`. This option shouldn't be necessary, but it doesn't hurt.